+++ instrumentation/opentelemetry-instrumentation-flask/src/opentelemetry/instrumentation/flask/__init__.py
@@ -110,9 +110,13 @@ def _before_request():
         return
 
     environ = flask.request.environ
-    span_name = flask.request.endpoint or otel_wsgi.get_default_span_name(
-        environ
-    )
+    span_name = None
+    try:
+        span_name = flask.request.url_rule.rule
+    except AttributeError:
+        pass
+    if span_name is None:
+        span_name = otel_wsgi.get_default_span_name(environ)
     token = context.attach(
         propagators.extract(otel_wsgi.get_header_from_environ, environ)
     )